+++ packages/multiple-line/src/styles.ts
@@ -1,4 +1,13 @@
 import { css } from 'emotion';
+import tokens from '@contentful/forma-36-tokens';
+
+export const validationRow = css({
+  display: 'flex',
+  justifyContent: 'space-between',
+  fontSize: tokens.fontSizeM,
+  marginTop: tokens.spacingXs,
+  color: tokens.gray700
+});
 
 export const rightToLeft = css({
   direction: 'rtl'